backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML libraries
torch==2.1.1
transformers==4.36.0
torchaudio==2.1.1
librosa==0.10.1
dtaidistance==2.3.10
numpy==1.24.3
scikit-learn==1.3.2

# Model conversion and deployment
tensorflow==2.15.0
onnx==1.15.0
onnxruntime==1.16.3

# Audio processing
soundfile==0.12.1
pydub==0.25.1

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
